worktree: Separate integration tests into separate target
3s for tests -> 1.8s
